requirements.txt
requests==2.31.0
Django==4.2.1
python-dotenv==1.0.1 
pillow==10.0.0